Separate standalone notebook embed path  (#7269)
* Separate standalone notebook embed path and make non-deferred for now

* cleanup cruft
