+++ bokeh/core/templates.py
@@ -1,11 +1,12 @@
-''' Provide Jinja2 templates used by Bokeh to embed Bokeh models
-(e.g. plots, widgets, layouts) in various ways.
+''' Provide Jinja2 templates used by Bokeh to embed Bokeh documents and
+models in various ways.
 
 .. bokeh-jinja:: bokeh.core.templates.AUTOLOAD_JS
 .. bokeh-jinja:: bokeh.core.templates.AUTOLOAD_NB_JS
 .. bokeh-jinja:: bokeh.core.templates.AUTOLOAD_TAG
 .. bokeh-jinja:: bokeh.core.templates.CSS_RESOURCES
 .. bokeh-jinja:: bokeh.core.templates.DOC_JS
+.. bokeh-jinja:: bokeh.core.templates.DOC_NB_JS
 .. bokeh-jinja:: bokeh.core.templates.FILE
 .. bokeh-jinja:: bokeh.core.templates.JS_RESOURCES
 .. bokeh-jinja:: bokeh.core.templates.NOTEBOOK_LOAD
@@ -31,6 +32,8 @@ SCRIPT_TAG = _env.get_template("script_tag.html")
 PLOT_DIV = _env.get_template("plot_div.html")
 
 DOC_JS = _env.get_template("doc_js.js")
+
+DOC_NB_JS = _env.get_template("doc_nb_js.js")
 
 FILE = _env.get_template("file.html")
